Consolidate logic for annotation <-> tab assignment
Previously the logic for determining which tab an annotation should appear
in occurred both in `root-thread.js` where it was used to filter
annotations based on the selected tab, and in `widget-controller.js` where
it was used to switch to the tab containing a particular annotation.

Consolidate all the logic for deciding which tab an annotation should
appear in into a single module. This makes it easier to test that tab
assignment is correct with different combinations of feature flags and
annotation states and to add additional logic in future.